added: loading bar to App.jsx
